requirements.txt
semantic-kernel[azure]
openai
python-dotenv
chainlit
aiohttp
azure-ai-projects 
azure-identity 
